+++ README.md
@@ -11,7 +11,15 @@ $ pip install git+https://github.com/tbongkim03/mov.git
 ```bash
 $ git clone <URL>
 $ cd <DIR>
+$ source .venv/bin/activate
 $ # option
 $ pdm venv create
-$ source .venv/bin/activate
+```
+
+### setting env
+``bash
+cat ~/.zshrc | tail -n 3
+
+# MYENV
+export MOVIE_API_KEY="<key>"
 ```
