+++ pyproject.toml
@@ -1,14 +1,15 @@
 [project]
 name = "mov"
-version = "0.2/api"
+version = "0.2.0"
 description = "Default template for PDM package"
 authors = [
     {name = "tbongkim03", email = "tbongkim03@gmial.com"},
 ]
 dependencies = [
     "requests>=2.32.3",
+    "pandas>=2.2.2",
 ]
-requires-python = "==3.11.*"
+requires-python = ">=3.9"
 readme = "README.md"
 license = {text = "MIT"}
