--- src/mov/api/test_call.py
@@ -1,25 +1,0 @@
-from mov.api.call import gen_url, req, get_key, req2dataframe
-
-def test_비밀키숨기기():
-    key = get_key()
-    assert "key"
-
-def test_유알엘테스트():
-    url = gen_url()
-    assert "http" in url
-
-def test_req():
-    code, data = req()
-    
-    assert code == 200
-
-    code, data = req('20230101')
-    assert code == 200
-
-def test_r2d():
-    l = req2dataframe() 
-    assert len(l) > 0
-    v = l[0]
-    assert 'rnum' in v.keys()
-    assert v['rnum'] == '1'
-    #assert False
